Neue Datenbankeinbindung finalisiert.
- improves #40
