+++ src/ui/components/sensors/index.js
@@ -1,2 +1,1 @@
 export { default as Sensor } from './_sensor.js';
-export { default as DashboardSensor } from './_dashboardSensor.js';
